% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{make_square_grid}
\alias{make_square_grid}
\title{Create a square grid}
\usage{
make_square_grid(nrow, ncol = nrow)
}
\arguments{
\item{nrow}{the number of rows in the grid.}

\item{ncol}{defaults to \code{nrow}. The number columns in the grid.}
}
\value{
An \code{sfc} object by \code{sf} package.
}
\description{
Creates a square grid with \code{ncol} and \code{nrow} dimensions.
}
\examples{
make_square_grid(3, 2)
}
