% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_durbin}
\alias{sim_durbin}
\title{Simulate the Spatial Durbin Model}
\usage{
sim_durbin(u, xb, wxg, listw, rho = 0.5)
}
\description{
Simulate the Spatial Durbin Model
}
\references{
\href{https://pysal.org/spreg/generated/spreg.dgp.dgp_spdurbin.html#spreg.dgp.dgp_spdurbin}{\code{spreg.dgp.dgp_spdurbin}}
}
