% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_error}
\alias{sim_error}
\title{Simulate Spatial Error Model}
\usage{
sim_error(u, listw, lambda = 0.5, model = c("sar", "ma"))
}
\description{
Simulate Spatial Error Model
}
\references{
See \href{https://pysal.org/spreg/generated/spreg.dgp.dgp_errproc.html#spreg.dgp.dgp_errproc}{\code{spreg.dgp.dgp_errproc}}
}
