% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_gns}
\alias{sim_gns}
\title{Simulate General Nested Model}
\usage{
sim_gns(u, xb, wxg, listw, rho = 0.5, lambda = 0.2, model = c("sar", "ma"))
}
\description{
Simulate General Nested Model
}
\references{
\href{https://pysal.org/spreg/generated/spreg.dgp.dgp_gns.html#spreg.dgp.dgp_gns}{\code{spreg.dgp.dgp_gns}}
}
