% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_mess}
\alias{sim_mess}
\title{Simiulate Matrix Exponential Spatial Lag Model}
\usage{
sim_mess(u, xb, listw, rho = 0.5)
}
\arguments{
\item{u}{an error vector}

\item{xb}{predicted x values as calculated by \code{make_xb()}}

\item{listw}{a \code{listw} object generated with \code{sim_grid_listw()}.}

\item{rho}{the spatial autoregressive coefficient for the spatially lagged dependent variable.}
}
\value{
A numeric vector
}
\description{
Simiulate Matrix Exponential Spatial Lag Model
}
\references{
\href{https://pysal.org/spreg/_modules/spreg/dgp.html#dgp_mess}{\code{dgp_mess}}
}
