% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_ols}
\alias{sim_ols}
\title{Simulate OLS}
\usage{
sim_ols(u, xb)
}
\description{
Simulate OLS
}
\examples{
u = make_x(50, method = "normal")
x = make_x(50)
xb = make_xb(x, c(1,2))
y <- dgp_ols(u, xb)
mod <- lm(y ~ x[[1]])
summary(mod)
}
\references{
\href{https://pysal.org/spreg/generated/spreg.dgp.dgp_ols.html#spreg.dgp.dgp_ols}{\code{spreg.dgp.dgp_ols}}
}
