% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_sarar}
\alias{sim_sarar}
\title{Simulate the Spatial Autoregressive Model with Autoregressive Errors}
\usage{
sim_sarar(u, xb, listw, rho = 0.5, lambda = 0.2, model = c("sar", "ma"))
}
\description{
Generate \code{y} values for the "combo" / SARAR / SAC model.
}
\references{
\href{https://pysal.org/spreg/generated/spreg.dgp.dgp_lagerr.html#spreg.dgp.dgp_lagerr}{\code{spreg.dgp.dgp_lagerr}}
}
