% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_sem}
\alias{sim_sem}
\title{Simulate Spatial Error Model (SEM)}
\usage{
sim_sem(u, xb, listw, lambda = 0.5, model = c("sar", "ma"))
}
\description{
Simulate Spatial Error Model (SEM)
}
\references{
\href{https://pysal.org/spreg/generated/spreg.dgp.dgp_sperror.html#spreg.dgp.dgp_sperror}{\code{spreg.dgp.dgp_sperror}}
}
