% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_slx_error}
\alias{sim_slx_error}
\title{Simulate Spatially Lagged X Error Model}
\usage{
sim_slx_error(u, xb, wxg, lambda = 0.5, model = c("sar", "ma"))
}
\description{
Simulate Spatially Lagged X Error Model
}
\references{
\href{https://pysal.org/spreg/generated/spreg.dgp.dgp_slxerror.html#spreg.dgp.dgp_slxerror}{\code{spreg.dgp.dgp_slxerror}}
}
