% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{make_wx}
\alias{make_wx}
\title{Create Spatial Lags of variables}
\usage{
make_wx(x, listw, order = NULL)
}
\arguments{
\item{order}{unused.}
}
\value{
A \code{data.frame} of the spatially lagged variables.
}
\description{
Given a dataframe of numeric values and a spatial weights matrix, calculate the spatial lag of each variable.
}
