% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sim_slx}
\alias{sim_slx}
\title{Simulate Spatially Lagged X (SLX) model}
\usage{
sim_slx(u, xb, wxg)
}
\description{
Simulate Spatially Lagged X (SLX) model
}
\examples{
n <- 100
grid <- make_grid(sqrt(n), sqrt(n))
listw <- spdep::nb2listw(spdep::poly2nb(grid))
u <- make_error(n, method = "normal")
x <- make_x(n, method = "uniform")
xb <- make_xb(x, c(1, 2))
wx <- make_wx(x, listw)
wxg <- make_wxg(wx, c(2, 2))
y <- dgp_slx(u, xb, wxg)
df <- data.frame(y, x)
spatialreg::lmSLX(y ~ ., data = df, listw = listw) 
}
\references{
\href{https://pysal.org/spreg/generated/spreg.dgp.dgp_slx.html#spreg.dgp.dgp_slx}{\code{spreg.dgp.dgp_slx}}
}
